assignment_lmwn/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
fastapi==0.109.0
FastAPI-SQLAlchemy==0.2.1
Flask==3.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
locust==2.20.1
Mako==1.3.0
MarkupSafe==2.1.4
msgpack==1.0.7
numpy==1.26.3
pandas==2.2.0
psutil==5.9.8
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pygeodesic==0.1.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
roundrobin==0.0.4
scikit-learn==1.4.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
